Fixing average movement calculation.

+++ Sensor/App.xaml.cs
@@ -287,7 +287,7 @@ namespace Sensor
 							MaxLight = this.maxLight,
 							MaxLightAt = this.maxLightAt,
 							AvgLight = (this.nrTerms == 0 ? (double?)null : this.sumLight / this.nrTerms),
-							AvgMovement = (this.nrTerms == 0 ? (double?)null : this.sumMovement / this.nrTerms)
+							AvgMovement = (this.nrTerms == 0 ? (double?)null : ((double)this.sumMovement) / this.nrTerms)
 						};
 
 						await Database.Insert(Rec);
